Update Main.kt
Updated line 71 so that the action "free" is actually tested and not just taken for granted if the action is not "mine". This eliminates some of the risks with typos when writing the command. If something else that "mine" or "free" is typed for the command nothing will happen and the player is just asked again.